# backend/app/services/migration_advisor/migration_advisor/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Cloud Migration Advisor API
  description: |
    Comprehensive API for the Cloud Migration Advisor platform that guides organizations through
    their cloud migration journey from initial assessment through provider selection, migration
    execution, and post-migration resource organization.
    
    ## Features
    - Migration project management and assessment
    - Workload and requirements analysis
    - ML-powered cloud provider recommendations
    - Migration planning and execution tracking
    - Automated resource discovery and organization
    - Multi-dimensional resource management
    - Post-migration FinOps integration
    
    ## Authentication
    All endpoints require authentication using Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
  version: 1.0.0
  contact:
    name: Cloud Migration Advisor Support
    email: support@cloudmigration.example.com
  license:
    name: Proprietary
servers:
  - url: https://api.cloudmigration.example.com/v1
    description: Production server
  - url: https://staging-api.cloudmigration.example.com/v1
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
...
security:
  - BearerAuth: []
paths:
...
  /api/migrations/{project_id}/integration/finops:
    post:
      tags:
        - Post-Migration Integration
      summary: Integrate with FinOps platform
      description: |
        Configure cost tracking, budget alerts, and optimization features.
        
        **Requirements:** 7.1, 7.2
      operationId: integrateWithFinOps
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enable_cost_tracking:
                  type: boolean
                  default: true
                enable_budget_alerts:
                  type: boolean
                  default: true
                enable_waste_detection:
                  type: boolean
                  default: true
                enable_optimization:
                  type: boolean
                  default: true
                cost_allocation_method:
                  type: string
                  default: "proportional"
            example:
              enable_cost_tracking: true
              enable_budget_alerts: true
              enable_waste_detection: true
              enable_optimization: true
              cost_allocation_method: "proportional"
      responses:
        '200':
          description: FinOps integration completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  integration_id:
                    type: string
                  status:
                    type: string
                  cost_tracking_enabled:
                    type: boolean
                  budget_alerts_configured:
                    type: integer
                  optimization_features_enabled:
                    type: array
                    items:
                      type: string
                  integration_timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid request
        '404':
          description: Project not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/migrations/{project_id}/integration/baselines:
    post:
      tags:
        - Post-Migration Integration
      summary: Capture baseline metrics
      description: |
        Capture initial cost and performance metrics after migration.
        
        **Requirements:** 7.3
      operationId: captureBaselineMetrics
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                capture_cost_data:
                  type: boolean
                  default: true
                capture_performance_data:
                  type: boolean
                  default: true
                capture_utilization_data:
                  type: boolean
                  default: true
                baseline_period_days:
                  type: integer
                  default: 7
                  minimum: 1
                  maximum: 30
            example:
              capture_cost_data: true
              capture_performance_data: true
              capture_utilization_data: true
              baseline_period_days: 7
      responses:
        '200':
          description: Baseline metrics captured
          content:
            application/json:
              schema:
                type: object
                properties:
                  baseline_id:
                    type: string
                  capture_date:
                    type: string
                    format: date-time
                  total_monthly_cost:
                    type: number
                    format: float
                  cost_by_service:
                    type: object
                    additionalProperties:
                      type: number
                  cost_by_team:
                    type: object
                    additionalProperties:
                      type: number
                  cost_by_project:
                    type: object
                    additionalProperties:
                      type: number
                  cost_by_environment:
                    type: object
                    additionalProperties:
                      type: number
                  resource_count:
                    type: integer
                  resource_count_by_type:
                    type: object
                    additionalProperties:
                      type: integer
        '400':
          description: Invalid request
        '404':
          description: Project not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
          
    get:
      tags:
        - Post-Migration Integration
      summary: Get baseline metrics
      description: |
        Retrieve baseline metrics for a migration project.
        
        **Requirements:** 7.3
      operationId: getBaselineMetrics
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      responses:
        '200':
          description: Baseline metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  baseline_id:
                    type: string
                  capture_date:
                    type: string
                    format: date-time
                  total_monthly_cost:
                    type: number
                  cost_by_service:
                    type: object
                  cost_by_team:
                    type: object
                  cost_by_project:
                    type: object
                  cost_by_environment:
                    type: object
                  resource_count:
                    type: integer
                  resource_count_by_type:
                    type: object
        '404':
          description: Baseline metrics not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/migrations/{project_id}/reports/final:
    get:
      tags:
        - Post-Migration Integration
      summary: Get migration report
      description: |
        Generate and retrieve comprehensive migration report with costs, timeline analysis,
        success metrics, lessons learned, and optimization opportunities.
        
        **Requirements:** 7.5
      operationId: getMigrationReport
      parameters:
        - name: project_id
          in: path
          required: true
          description: Migration project ID
          schema:
            type: string
      responses:
        '200':
          description: Migration report
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    type: string
                  report_date:
                    type: string
                    format: date-time
                  start_date:
                    type: string
                    format: date-time
                  completion_date:
                    type: string
                    format: date-time
                  actual_duration_days:
                    type: integer
                  planned_duration_days:
                    type: integer
                  total_cost:
                    type: number
                    format: float
                  budgeted_cost:
                    type: number
                    format: float
                  cost_variance_percentage:
                    type: number
                    format: float
                  resources_migrated:
                    type: integer
                  success_rate:
                    type: number
                    format: float
                  lessons_learned:
                    type: array
                    items:
                      type: string
                  optimization_opportunities:
                    type: array
                    items:
                      type: object
                  cost_analysis:
                    type: object
                  timeline_analysis:
                    type: object
        '404':
          description: Project not found
        '401':
          description: Unauthorized
        '500':
          description: Internal server error